requirements.txt
flask~=3.1.0
werkzeug~=3.1.3
python-dotenv~=1.0.1
numpy~=2.0.2